pyproject.toml
[tool.poetry]
name = "aggrec"
version = "1.0.0"
description = "DNS TAPIR Aggregate Receiver"
authors = ["Jakob Schlyter <jakob@kirei.se>"]
readme = "README.md"

[tool.poetry.scripts]
aggrec_server = "aggrec.server:main"
aggrec_client = "aggrec.client:main"

[tool.poetry.dependencies]
python = "^3.12"
mongoengine = "^0.29.0"
http-message-signatures = "^0.5.0"
requests = "^2.30.0"
fastapi = "^0.114.0"
uvicorn = "^0.30.1"
pydantic-settings = "^2.3.4"
werkzeug = "^3.0.4"
boto3 = "^1.26.133"
aiobotocore = "^2.13.3"
aiomqtt = "^2.2.0"
jsonformatter = "^0.3.2"
pendulum = "^3"
pyyaml = "^6.0.1"
opentelemetry-instrumentation-fastapi = "^0.48b0"
opentelemetry-exporter-otlp = "^1.27.0"
opentelemetry-instrumentation-botocore = "^0.48b0"
opentelemetry-instrumentation-pymongo = "^0.48b0"
http-sf = "^1.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.6.3"
pytest-ruff = "^0.4.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["E501", "I001"]

[tool.ruff.lint.isort]
force-sort-within-sections = false
known-first-party = [
    "aggrec",
]
extra-standard-library = [
    "tomllib",
]
combine-as-imports = true
split-on-trailing-comma = false
